requirements.txt
Flask==2.3.2
tensorflow==2.12.0
google-cloud-storage==2.9.0
Pillow==9.5.0
gunicorn==20.1.0
flask-cors==4.0.0